@@ -119,26 +96,26 @@
     "@babel/preset-env": "^7.14.5",
     "@babel/preset-typescript": "^7.14.5",
     "@rollup/plugin-babel": "^6.0.0",
-    "@rollup/plugin-commonjs": "^24.0.0",
     "@rollup/plugin-node-resolve": "^15.0.0",
     "@rollup/plugin-terser": "^0.4.0",
     "@types/jest": "^29.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.17.0",
     "eslint": "^8.0.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.0",
     "jest": "^29.0.0",
     "jest-environment-jsdom": "^29.0.0",
-    "prettier": "^2.3.2",
-    "release-it": "^15.0.0",
-    "rollup": "^3.0.0",
+    "prettier": "^3.0.0",
+    "release-it": "^17.0.0",
+    "rollup": "4.9",
     "rollup-plugin-license": "^3.0.0",
-    "typedoc": "^0.23.10",
+    "typedoc": "^0.25.6",
     "typedoc-plugin-markdown": "^3.9.0",
-    "typescript": "^4.2.4"
+    "typescript": "5.3"
   },
   "dependencies": {
     "@babel/plugin-transform-runtime": "^7.14.5",
+    "@babel/runtime": "^7.23.7",
     "core-js": "^3.14.0"
   }
 }
