@@ -2,7 +2,6 @@ import { babel } from "@rollup/plugin-babel";
 import terser from "@rollup/plugin-terser";
 import { nodeResolve } from "@rollup/plugin-node-resolve";
 import license from "rollup-plugin-license";
-import commonjs from "@rollup/plugin-commonjs";
 
 const modules = [
   "array",
@@ -14,7 +13,7 @@ const modules = [
   "guard",
   "index",
   "iterable",
-  "map",
+  "number",
   "object",
   "regexp",
   "string",
@@ -29,86 +28,25 @@ const extensions = [".ts", ".js"];
 
 const banner = {
   content:
-    "Shed | Copyright (C) 2023 Peter Kröner | peter@peterkroener.de | MIT",
+    "Shed | Copyright (C) 2024 Peter Kröner | peter@peterkroener.de | MIT",
   commentStyle: "ignored",
 };
 
-const moduleBaseConfig = {
-  external: [/@sirpepe\/shed/, /@babel\/runtime/, /core-js/],
+export default modules.map((module) => ({
+  input: `./src/${module}.ts`,
   plugins: [
     nodeResolve({ extensions }),
     babel({
       extensions,
       babelHelpers: "runtime",
       exclude: "node_modules/**",
-      presets: [
-        [
-          "@babel/preset-env",
-          {
-            targets: "defaults,not ie 11",
-            useBuiltIns: "usage",
-            corejs: "3.8",
-          },
-        ],
-        "@babel/preset-typescript",
-      ],
+      presets: ["@babel/preset-env", "@babel/preset-typescript"],
       plugins: ["@babel/plugin-transform-runtime"],
     }),
   ],
-};
-
-const browserBaseConfig = {
-  plugins: [
-    nodeResolve({ extensions }),
-    commonjs(),
-    babel({
-      extensions,
-      babelHelpers: "inline",
-      exclude: "node_modules/**",
-      presets: [
-        [
-          "@babel/preset-env",
-          {
-            targets: "defaults,not ie 11",
-            useBuiltIns: "usage",
-            corejs: "3.8",
-          },
-        ],
-        "@babel/preset-typescript",
-      ],
-    }),
-  ],
-};
-
-export default [
-  ...modules.flatMap((module) => [
-    {
-      input: `./src/${module}.ts`,
-      output: {
-        file: `dist/esm/${module}.js`,
-        format: "esm",
-        plugins: [license({ banner })],
-      },
-      ...moduleBaseConfig,
-    },
-    {
-      input: `./src/${module}.ts`,
-      output: {
-        file: `dist/cjs/${module}.js`,
-        format: "cjs",
-        plugins: [license({ banner })],
-      },
-      ...moduleBaseConfig,
-    },
-  ]),
-  {
-    input: "./src/index.ts",
-    output: {
-      file: `dist/min/index.js`,
-      format: "umd",
-      name: "_",
-      plugins: [terser(), license({ banner })],
-    },
-    ...browserBaseConfig,
+  output: {
+    file: `dist/${module}.js`,
+    format: "esm",
+    plugins: [terser(), license({ banner })],
   },
-];
+}));
