@@ -1,6 +1,6 @@
 /**
  * A sorted array-like data structure. Use this module by importing from
- * `@sirpepe/shed/SortedArray` or `@sirpepe/shed`.
+ * `@sirpepe/shed/SortedArray`.
  *
  * @module
  */
