@@ -1,6 +1,5 @@
 /**
- * Error utilities. Use this module by importing from `@sirpepe/shed/error`
- * or `@sirpepe/shed`.
+ * Error utilities. Use this module by importing from `@sirpepe/shed/error`.
  *
  * @module
  */
