@@ -1,6 +1,6 @@
 /**
  * Collection of type guards. Use this module by importing from
- * `@sirpepe/shed/guard` or `@sirpepe/shed`.
+ * `@sirpepe/shed/guard`.
  *
  * @module
  */
