@@ -6,7 +6,6 @@ export * as FRDY from "./FRDY";
 export { debounce, debounceRaf, identity, noop } from "./function";
 export { is, isNot } from "./guard";
 export { groupBy, mapBy, partition } from "./iterable";
-export { isMap } from "./map";
 export { isReal, roundFloat, ceilFloat, floorFloat, stringify } from "./number";
 export {
   omitted,
@@ -18,7 +17,7 @@ export {
   select,
 } from "./object";
 export { isRegExp, escapeRegExpString } from "./regexp";
-export { isSet, addAll } from "./set";
+export { addAll } from "./set";
 export { SortedArray } from "./SortedArray";
 export { asc, desc } from "./sorting";
 export { uppercaseFirst, lowercaseFirst, length } from "./string";
