@@ -1,6 +1,6 @@
 /**
  * Utilities for iterable objects. Use this module by importing from from
- * `@sirpepe/shed/iterable` or `@sirpepe/shed`.
+ * `@sirpepe/shed/iterable`.
  *
  * @module
  */
