@@ -1,6 +1,6 @@
 /**
  * Callback functions for use with Array.prototype.sort(). Use this module by
- * importing from `@sirpepe/shed/string` or `@sirpepe/shed`.
+ * importing from `@sirpepe/shed/sorting`.
  *
  * @module
  */
