@@ -1,15 +1,6 @@
-import { addAll, isSet } from "../src/set";
+import { addAll } from "../src/set";
 
 describe("set", () => {
-  describe("isSet()", () => {
-    test("identifies maps", () => {
-      expect(isSet(new Set())).toEqual(true);
-      expect(isSet(new Map())).toEqual(false);
-      expect(isSet(undefined)).toEqual(false);
-      expect(isSet("test")).toEqual(false);
-    });
-  });
-
   describe("addAll()", () => {
     test("adds multiple values at once", () => {
       const set = new Set();
